requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
anthropic>=0.39.0
openai>=1.50.0
supabase>=2.9.0
python-dotenv>=1.0.0
